# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API server for SMM-auto application",
        "title": "SMM-Auto API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "paths": {
...
        "/channel/": {
            "get": {
                "description": "Returns an array of Channel objects",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "channel"
                ],
                "summary": "Get all channels",
                "operationId": "GetAllChannels",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.GetAllChannelsResponse"
                        }
                    },
                    "400": {
                        "description": "Error identifying user",
                        "schema": {
                            "$ref": "#/definitions/jsonHelper.ApiError"
                        }
                    },
                    "417": {
                        "description": "Error identifying user",
                        "schema": {
                            "$ref": "#/definitions/jsonHelper.ApiError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/jsonHelper.ApiError"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/jsonHelper.ApiError"
                        }
                    }
                }
            }
        },
        "/channel/add": {
            "post": {
                "description": "Remove channel from user's list",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "channel"
                ],
                "summary": "Delete channel",
                "operationId": "DeleteChannel",
                "parameters": [
                    {
                        "type": "string",
                        "description": "channel ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Error identifying user",
                        "schema": {
                            "$ref": "#/definitions/jsonHelper.ApiError"
                        }
                    },
                    "403": {
                        "description": "User does not have access to this channel",
                        "schema": {
                            "$ref": "#/definitions/jsonHelper.ApiError"
                        }
                    },
                    "417": {
                        "description": "Error identifying user",
                        "schema": {
                            "$ref": "#/definitions/jsonHelper.ApiError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/jsonHelper.ApiError"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/jsonHelper.ApiError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controllers.GetAllChannelsResponse": {
            "type": "object",
            "properties": {
                "channels": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Channel"
                    }
                }
            }
        },
...
        "jsonHelper.ApiError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "models.Channel": {
            "type": "object",
            "properties": {
                "assignedBotToken": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}